+++ pkg/validator/others.go
@@ -12,9 +12,14 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
+type SourceTenantMetrics struct {
+	Regexp      string `yaml:"regexp"`
+	Description string `yaml:"description"`
+}
+
 func newHasSourceTenantsForMetrics(paramsConfig yaml.Node) (Validator, error) {
 	params := struct {
-		SourceTenants map[string]string `yaml:"sourceTenants"`
+		SourceTenants map[string]SourceTenantMetrics `yaml:"sourceTenants"`
 	}{}
 	if err := paramsConfig.Decode(&params); err != nil {
 		return nil, err
@@ -22,27 +27,35 @@ func newHasSourceTenantsForMetrics(paramsConfig yaml.Node) (Validator, error) {
 	if params.SourceTenants == nil || len(params.SourceTenants) == 0 {
 		return nil, fmt.Errorf("sourceTenants metrics mapping needs to be set")
 	}
-	validator := hasSourceTenantsForMetrics{sourceTenants: map[string]*regexp.Regexp{}}
-	for tenant, metricsRegexp := range params.SourceTenants {
-		compiledRegexp, err := regexp.Compile("^" + metricsRegexp + "$")
+	validator := hasSourceTenantsForMetrics{sourceTenants: map[string]tenantMetrics{}}
+	for tenant, metrics := range params.SourceTenants {
+		compiledRegexp, err := regexp.Compile("^" + metrics.Regexp + "$")
 		if err != nil {
-			return nil, fmt.Errorf("invalid metric name regexp: %s", metricsRegexp)
+			return nil, fmt.Errorf("invalid metric name regexp: %s", metrics.Regexp)
 		}
-		validator.sourceTenants[tenant] = compiledRegexp
+		validator.sourceTenants[tenant] = tenantMetrics{
+			regexp:      compiledRegexp,
+			description: metrics.Description,
+		}
 	}
 	return &validator, nil
 }
 
+type tenantMetrics struct {
+	regexp      *regexp.Regexp
+	description string
+}
+
 type hasSourceTenantsForMetrics struct {
-	sourceTenants map[string]*regexp.Regexp
+	sourceTenants map[string]tenantMetrics
 }
 
 func (h hasSourceTenantsForMetrics) String() string {
 	tenantStrings := []string{}
 	for tenant, metricsRegexp := range h.sourceTenants {
-		tenantStrings = append(tenantStrings, fmt.Sprintf("`%s`:`%s`", tenant, metricsRegexp.String()))
+		tenantStrings = append(tenantStrings, fmt.Sprintf("`%s`:   `%s` (%s)", tenant, metricsRegexp.regexp.String(), metricsRegexp.description))
 	}
-	return fmt.Sprintf("rule group, the rule belongs to, has the required `source_tenants` configured, according to the mapping of metric names to tenants: %s", strings.Join(tenantStrings, ", "))
+	return fmt.Sprintf("rule group, the rule belongs to, has the required `source_tenants` configured, according to the mapping of metric names to tenants: \n        %s", strings.Join(tenantStrings, "\n        "))
 }
 
 func (h hasSourceTenantsForMetrics) Validate(group unmarshaler.RuleGroup, rule rulefmt.Rule, _ *prometheus.Client) []error {
@@ -54,7 +67,7 @@ func (h hasSourceTenantsForMetrics) Validate(group unmarshaler.RuleGroup, rule r
 	}
 	for _, usedMetric := range usedMetrics {
 		for tenant, metricsRegexp := range h.sourceTenants {
-			if metricsRegexp.MatchString(usedMetric.Name) && !slices.Contains(group.SourceTenants, tenant) {
+			if metricsRegexp.regexp.MatchString(usedMetric.Name) && !slices.Contains(group.SourceTenants, tenant) {
 				errs = append(errs, fmt.Errorf("rule uses metric `%s` of the tenant `%s` tenant, you should set the tenant in the groups source_tenants settings", tenant, usedMetric.Name))
 			}
 		}